#616: fixed menus tabs
