+++ messages/en.json
@@ -843,6 +843,7 @@
       "Validators": "Validators",
       "Nodes": "Nodes",
       "Networks": "Networks",
+      "Mining Pools": "Mining Pools",
       "Ecosystems": "Ecosystems",
       "Metrics": "Metrics"
     },
