+++ messages/pt.json
@@ -842,6 +842,7 @@
     "Tabs": {
       "Validators": "Validadores",
       "Networks": "Redes",
+      "Mining Pools": "Pools de Mineração",
       "Metrics": "Métricos",
       "Ecosystems": "Ecossistemas",
       "Nodes": "Nodes"
