+++ messages/ru.json
@@ -842,6 +842,7 @@
     "Tabs": {
       "Validators": "Валидаторы",
       "Networks": "Сети",
+      "Mining Pools": "Майнинг-пулы",
       "Metrics": "Метрики",
       "Ecosystems": "Экосистемы",
       "Nodes": "Ноды"
