+++ src/app/[locale]/networks/[name]/mining-pools/network-mining-pools-table/network-mining-pools.tsx
@@ -60,7 +60,7 @@ const NetworkMiningPools: FC<OwnProps> = async ({ chainName, window, sort }) => 
 
   return (
     <div>
-      <div className="mt-6 flex items-center gap-4">
+      <div className="mt-6 flex flex-row items-end justify-end">
         {windowOptions.length > 1 && <HashrateWindowSelector current={safeWindow} options={windowOptions} />}
       </div>
       <BaseTable className="my-4">
