+++ src/app/[locale]/p2pchat/page.tsx
@@ -5,7 +5,7 @@ import CollapsiblePageHeader from '@/app/validators/collapsible-page-header';
 import PageHeaderVisibilityWrapper from '@/components/common/page-header-visibility-wrapper';
 import PageTitle from '@/components/common/page-title';
 import TabList from '@/components/common/tabs/tab-list';
-import { mainTabs } from '@/components/common/tabs/tabs-data';
+import { toolsTabsHorizontal } from '@/components/common/tabs/tabs-data';
 import { NextPageWithLocale } from '@/i18n';
 
 export const dynamic = 'force-dynamic';
@@ -16,7 +16,7 @@ const RumorsPage: NextPageWithLocale = async ({ params: { locale } }) => {
   return (
     <div className="flex flex-grow flex-col">
       <PageHeaderVisibilityWrapper>
-        <TabList page="HomePage" tabs={mainTabs} />
+        <TabList page="HomePage" tabs={toolsTabsHorizontal} />
       </PageHeaderVisibilityWrapper>
       <CollapsiblePageHeader description={t('description')}>
         <PageTitle text={t('title')} />
